agents-stage-live2d-vrm3d-server/pyproject.toml
[project]
name = "live2d-assistant-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp",
    "ollama",
    "openai",
    "python-dotenv",
    "fastapi[standard]",
    "uvicorn",
    "pydantic",
    "requests",
    "selenium",
    "beautifulsoup4",
    "torch",
    "torchaudio",
    "python-dotenv",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langgraph",
    "ChatTTS",
    "duckduckgo-search",
    "langchain-mcp-adapters",
    "pymilvus",
    "sentence-transformers",
    "jieba",
    "PyMuPDF",
    "python-docx",
    "pytesseract",
    "Pillow",
    "transformers",
    "docx2txt",
    "rapidocr-onnxruntime>=1.4.4",
    "python-jose[cryptography]>=3.3.0",
    "fastapi-sso>=0.16.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
